Add files via upload
I was able to fix the TODO's that i had on the last update, but this time, I tried to make a checkwin, but have to think though how to do it first! Overall it is basically done, So I am going to start on a new Oregon trail mock game. I make a lot of games because that way I have something to compare it too and make sure my code works well.